Added sample
