+++ Word-document/Identify-Word-document-is-Encrypted-or-Protected/.NET/Identify-Word-document-is-Encrypted-or-Protected.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.14.37411.7 d17.14
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Identify-Word-document-is-Encrypted-or-Protected", "Identify-Word-document-is-Encrypted-or-Protected\Identify-Word-document-is-Encrypted-or-Protected.csproj", "{14F884CF-1AC7-B41A-C300-C3991FDE4A7F}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{14F884CF-1AC7-B41A-C300-C3991FDE4A7F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{14F884CF-1AC7-B41A-C300-C3991FDE4A7F}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{14F884CF-1AC7-B41A-C300-C3991FDE4A7F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{14F884CF-1AC7-B41A-C300-C3991FDE4A7F}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {F71E08D4-7F59-4261-8BD3-E1D09BE390C0}
+	EndGlobalSection
+EndGlobal
