+++ Word-document/Identify-Word-document-is-Encrypted-or-Protected/.NET/Identify-Word-document-is-Encrypted-or-Protected/Identify-Word-document-is-Encrypted-or-Protected.csproj
