--- Word-document/Identify-Word-document-is-Encrypted-or-not/.NET/Identify-Word-document-is-Encrypted-or-not.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.14.37012.4 d17.14
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Identify-Word-document-is-Encrypted-or-not", "Identify-Word-document-is-Encrypted-or-not\Identify-Word-document-is-Encrypted-or-not.csproj", "{54A935D7-C09A-697B-0D61-C7AAC43F2784}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{54A935D7-C09A-697B-0D61-C7AAC43F2784}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{54A935D7-C09A-697B-0D61-C7AAC43F2784}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{54A935D7-C09A-697B-0D61-C7AAC43F2784}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{54A935D7-C09A-697B-0D61-C7AAC43F2784}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {FCABE1DE-F4AA-4316-AEF1-3C5C985F9610}
-	EndGlobalSection
-EndGlobal
